% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/election_2006.R
\docType{data}
\name{election_2006}
\alias{election_2006}
\title{Presidential electoral results 2006.}
\format{A data frame with 130777 rows and 18 columns:
\describe{
  \item{state_code, state_name, state_abbr}{Character variables indicating
  the state corresponding to the polling station. State code's follow INEGI's
  standard}
  \item{polling_id}{Numeric identifier of the polling station}
  \item{stratum}{Stratum for quickcount sampling design}
  \item{distrito_fed}{Fedral district, the districts are ploygons that
  partition each state, they are defined by INE's cartography}
  \item{section}{The section is the miminal electoral geographical unit that
  contains the polling station}
  \item{region}{The region (center, north, ...) to which the state belongs}
  \item{polling_type}{Type of polling station, Basic (B) is the
  basic and most common type, Contigua (C) these polling stations arise when
  the nominal list surpases 750 people, Extraoridinaria (E) these are
  installed when people can not reach their planned station and Especial (S)
  polling stations are where people can vote outside their asigned polling
  station}
  \item{section_type}{Section type (in 2017), whether the section of the
  station is rural (R), urban (U) or mixed (M)}
  \item{pri_pvem}{Number of votes favoring the parties PRI and/or PVEM}
  \item{pan}{Number of votes favoring PAN}
  \item{panal}{Number of votes favoring Partido Nueva Alianza}
  \item{prd_pt_conv}{Number of votes favoring the parties PRD, PT,
  Convergencia}
  \item{psd}{Number of votes favoring PSD}
  \item{otros}{Number of votes that do not favor any of the parties (null,
  non-registered candidates)}
  \item{total}{Total number of votes registered}
  \item{nominal_list}{Nominal list, number of people registered to vote}
}}
\source{
\url{https://cartografia.ife.org.mx}
}
\usage{
election_2006
}
\description{
A dataset containing the election results for president 2006, each row
corresponds to a polling station.
}
\keyword{datasets}
