% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm25_2019.R
\docType{data}
\name{pm25_2019}
\alias{pm25_2019}
\title{PM2.5 pollutant measurements in Mexico City}
\format{A data frame with 5088 rows and 26 columns:
\describe{
  \item{date}{Date when the measurment was taken}
  \item{hour}{Hour at which the measurment was taken, 1 to 24}
  \item{AJM, AJU,...,XAL}{Initials indication the monitoring station}
}}
\usage{
pm25_2019
}
\description{
A data set containing the measuremnts of PM2.5, particles smaller than 2.5
micrometers measured in micrograms per square meter, across 70 stations
within Mexico City for the first months of 2019.
}
\keyword{datasets}
