% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata_sample_2006.R
\docType{data}
\name{strata_sample_2006}
\alias{strata_sample_2006}
\title{Sample of 2006 presidential elections.}
\format{strata_sample_2006: A data frame with 300 rows and 3 columns, one
row per stratum:
\describe{
  \item{stratum}{stratum id}
  \item{n}{Number of polling stations in the sample}
  \item{N}{Total number of polling stations in the stratum}
}}
\usage{
strata_sample_2006
}
\description{
A dataset containing the strata information for the 2006 presidential
election sample `sample_2006`
}
\keyword{datasets}
